// source/apps/system/public/api/schemas/login.input.ts
/**
 * This file was automatically generated by source/platform/system/schemas/json2ts.ts
 * DO NOT MODIFY IT MANUALLY. Instead, modify the source JSON Schema file,
 * and run 'npm run schemas' to regenerate this file.
 * The source file has the same name as this one but has the extension .json
 */

export interface InputSystemPublicLogin {
  username: string;
  password: string;
  language?: string;
  push_notifications_token?: string;
}