// source/apps/system/public/api/schemas/login.output.ts
/**
 * This file was automatically generated by source/platform/system/schemas/json2ts.ts
 * DO NOT MODIFY IT MANUALLY. Instead, modify the source JSON Schema file,
 * and run 'npm run schemas' to regenerate this file.
 * The source file has the same name as this one but has the extension .json
 */

export interface OutputSystemPublicLogin {
  fullname: string;
  email: string;
  token: string;
  timezone: string;
  i18n: {
    [k: string]: unknown;
  };
}